Aggiunta a configure.ac l'opzione --enable-debug-mode (in analogia a libsim) Aggiornati i files: arkimet_aree.dat e grb_chimere2netcdf.f90 Risolti bug nei sorgenti: seriet_utilities.f90, zoom_grib.f90, lm_levels2layers.f90, crea_surf_dat.f90 Ripuliti alcuni programmi dai warning in compilazione in debug-mode
git-svn-id: https://svn.smr.arpa.emr.it/metambdevel/ma_utils@81 465961ed-461c-4c0f-bd85-d0e70b234b9d

+++ arkimet/src/gacsv_scan.f90
@@ -21,7 +21,7 @@ PROGRAM gacsv_scan
 !      programma  da' un warning, e non sara' possibile convertire i dati
 !      in formato seriet.
 !
-!                                         Versione 1.3.1, Enrico 10/10/2013
+!                                         Versione 1.3.2, Enrico 17/10/2013
 !--------------------------------------------------------------------------
 
 USE file_utilities
@@ -165,8 +165,8 @@ DO k = 1,HUGE(0)
       npt = npt + 1
       IF (npt > maxpt) GOTO 9995
       pts_list(npt) = report%np
-      pts_lon(npt) = report%lon
-      pts_lat(npt) = report%lat
+      pts_lon(npt) = REAL(report%lon)
+      pts_lat(npt) = REAL(report%lat)
     ENDIF
 
 !   Controllo se ho trovato un nuovo var-liv
